+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # uppercut Changelog
 
-## [Unreleased]
+## [3.0.0] - 2026-09-05
 
 ### Added